(.#78) Add authenticator for Windows Phone
2fa.md did not contain any applications that work with Windows Phone,
despite there being a column dedicated to the OS.

To remedy this, Microsoft Authenticator has been linked to as an
application to fill this need.

images/1.8/windows-store.png has been added after compression. It is,
to my knowledge, the official icon to use since it was referenced in a
Microsoft blog post (http://bit.ly/1K7L1r3 ;
original host: blogs.windows.com).

Signed-off-by: Josh Harmon <me@joshharmon.me>